Fixed errors in services
