+++ lib/pages/spaces/tabletop/object_create_menu.dart
@@ -1,7 +1,7 @@
-import 'package:chat_interface/controller/spaces/tabletop/objects/tabletop_inventory.dart';
+import 'package:chat_interface/pages/spaces/tabletop/objects/tabletop_inventory.dart';
 import 'package:chat_interface/controller/spaces/tabletop/tabletop_controller.dart';
-import 'package:chat_interface/controller/spaces/tabletop/objects/tabletop_deck.dart';
-import 'package:chat_interface/controller/spaces/tabletop/objects/tabletop_text.dart';
+import 'package:chat_interface/pages/spaces/tabletop/objects/tabletop_deck.dart';
+import 'package:chat_interface/pages/spaces/tabletop/objects/tabletop_text.dart';
 import 'package:chat_interface/theme/ui/dialogs/window_base.dart';
 import 'package:chat_interface/util/vertical_spacing.dart';
 import 'package:flutter/material.dart';
