+++ lib/pages/spaces/tabletop/tabletop_page.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'dart:ui';
 
-import 'package:chat_interface/controller/spaces/tabletop/objects/tabletop_card.dart';
+import 'package:chat_interface/pages/spaces/tabletop/objects/tabletop_card.dart';
 import 'package:chat_interface/controller/spaces/tabletop/tabletop_controller.dart';
-import 'package:chat_interface/controller/spaces/tabletop/objects/tabletop_deck.dart';
+import 'package:chat_interface/pages/spaces/tabletop/objects/tabletop_deck.dart';
 import 'package:chat_interface/pages/settings/town/tabletop_settings.dart';
 import 'package:chat_interface/pages/settings/data/entities.dart';
 import 'package:chat_interface/pages/settings/data/settings_controller.dart';
