+++ lib/services/spaces/space_message_service.dart
@@ -1,0 +1,29 @@
+import 'package:chat_interface/controller/spaces/spaces_controller.dart';
+import 'package:chat_interface/services/connection/connection.dart';
+import 'package:chat_interface/services/spaces/space_message_provider.dart';
+import 'package:chat_interface/util/logging_framework.dart';
+
+class SpaceMessageService {
+  static void setupSpaceMessageListeners(Connector connector) {
+    // Listen for deletions
+    connector.listen("msg", (event) async {
+      // Make sure we're actually in a space right now
+      if (!SpacesController.connected.value || SpacesController.key == null) {
+        sendLog("WARNING: received space message even though not in space");
+        return;
+      }
+
+      // Unpack the message in a different isolate (to prevent lag)
+      final message = await SpacesMessageProvider.unpackMessageInIsolate(event.data["msg"]);
+
+      // Check if there are too many attachments
+      if (message.attachments.length > 5) {
+        sendLog("WARNING: invalid message, more than 5 attachments");
+        return;
+      }
+
+      // Tell the controller about the message
+      SpacesController.addMessage(message);
+    });
+  }
+}
